+++ project/brand_db/update_brand_db.py
@@ -4,7 +4,7 @@ import json
 import httpx
 import psycopg
 from tqdm.asyncio import tqdm
-from project.backend.app.core.settings import load_backend_env
+from project.backend.app.utils.settings import load_backend_env
 
 # Load environment variables
 load_backend_env()
